Move audit log modal into seperate file and access audit log data in there
